recommendation_system_service/requirements.txt
confluent_kafka
SQLAlchemy==2.0.20
asyncpg==0.27.0
geopy==2.3.0
python-dotenv==1.0.0
